Inline the iterator module.
When we make the iterator a more genuine iterator, we'll just end up
`use super::*`ing things, so we might as well inline it now.
